chore(refactor): delete unnecessary component
