+++ components/pages/Medic/Dashboard/index.jsx
@@ -1,11 +1,13 @@
 import React, { PureComponent } from 'react'
 import Header from '../../../shared/Header'
-import AppointmentListItem from '../../../shared/AppointmentListItem'
 import { Grid, Button, Loader, Dimmer, Image, Card } from 'semantic-ui-react'
 import styled from 'styled-components'
 import media from 'styled-media-query'
 import NewsApi from '../../../../utils/NewsAPI'
 import Link from 'next/link'
+import Avatar from '../../../shared/AppointmentAvatar'
+import ConsultationInfo from '../../../shared/AppointmentInfo'
+import ValueItem from '../../../shared/ValueItem'
 
 class DoctorDashboard extends PureComponent {
   state = {
@@ -46,7 +48,19 @@ class DoctorDashboard extends PureComponent {
                     Próximas Consultas
                   </Card.Header>
                   <Consultations>
-                    <AppointmentListItem appointment={appointment}/>
+                    <ConsultationRow>
+                      <AvatarContainer mobile={16} tablet={3} computer={2}>
+                        <Avatar src={ appointment.profile }/>
+                      </AvatarContainer>
+                      <ConsultationInfoContainer mobile={16} tablet={13} computer={14}>
+                        <Link as="" href=""><a>{ appointment.name }</a></Link>
+                        <ConsultationInfo>
+                          <ValueItem keyName="Fecha" value={ appointment.date } />
+                          <ValueItem keyName="Hora" value={ appointment.time } />
+                          <ValueItem keyName="Última cita" value={ appointment.previous } />
+                        </ConsultationInfo>
+                      </ConsultationInfoContainer>
+                    </ConsultationRow>
                   </Consultations>
                   <CenteredButton>
                     <MoreButton secondary className="small">Ver más</MoreButton>
@@ -198,4 +212,31 @@ const CenteredButton = styled.div`
 `
 const Consultations = styled.div`
   margin-bottom: 1rem;
+`
+const ConsultationRow = styled(Grid)`
+  >.column:not(.row):last-child {
+    padding-top: 1.5rem;
+  }
+`
+const AvatarContainer = styled(Grid.Column)`
+  text-align: center;
+
+  ${media.between('medium','1025px')`
+    [class*="two wide computer"]&&&.column {
+      width: 18.75%!important;
+    }
+  `}
+
+  ${media.lessThan('medium')`
+    &&&.column:not(.row) {
+      padding-bottom: 0;
+    }
+  `}
+`
+const ConsultationInfoContainer = styled(Grid.Column)`
+  ${media.between('medium','1025px')`
+    [class*="fourteen wide computer"]&&&.column {
+      width: 81.25%!important;
+    }
+  `}
 `
